fixed seeder
